test/testproject/requirements.txt
Pillow
django-autocomplete-light
django-tagging
sorl-thumbnail
lxml
swapper
django-cms
